refactor(review): improve API and reduce noise
Reduce need for pointers, remove leaky fs interactions types and
update tests.

Rinor albeit breaking changes to dsl package
