+++ examples/go-kit/README.md
@@ -89,7 +89,7 @@ Before you can run the consumer make sure the provider is
 [running](#running-the-provider).
 
 ```
-go run cmd/usersvc/main.go
+go run cmd/web/main.go
 ```
 
 Hit http://localhost:8081/ in your browser. You can use the username/password of
